Add Feature Policy feature for Trust Tokens
The experimental Trust Token API
(https://github.com/wicg/trust-token-api) comprises three operations:
issuance, redemption, and request signing. State for redemption and
signing is scoped by top-level origin and, as a consequence, we want
it to only be available to nested frames with permission from the main
frame. This change helps implement this behavior by adding a Feature
Policy feature for token redemption (and signing, which is closely
related) and passing the feature to URLLoaderFactoryParams, whence
the Trust Tokens logic in the network stack will be able to access
the state during request handling.

Bug: 1042962
Change-Id: I3369a18bdfd53538a47dcaffa79077a5df6966c9