Add tests for true computed value of transform interpolations
Create a set of interpation tests for transforms which read the true
computed value (as opposed to the resolved value returned by
getComputedStyle() which differs drastically in the case of transforms).
As there is currently no universal way to get this value,
each test comes in 3 versions:

Two variants using computedStyleMap().get(). toStrung() which is
supported by chrome. These also have suplicate versions where the zoom
property is set as it currently triggers some transform serialization
bugs in blink.

A version using commitStyles, which in implemented in both firefox and
chrome, but currently gives resolved values instead of computed ones
in chrome.

These assume the interperetation of computed value discussed in
https://github.com/w3c/csswg-drafts/issues/4869
which seems to be followed by both firefox and chrome.

Change-Id: I59afcc9f73906051212cc71570631509920ddc32