Fix web-exposed First Input Delay computation for desktop.
On desktop, when we click, the input event sequence would be pointerdown
,mousedown, pointerup, mouseup. The first input should be the
pointerdown. But our current FID computation would regard the mousedown
as the first input. And the input delay for the mousedown and
pointerdown can be different. This CL is to fix the web-exposed FID by
delaying reporting the mousedown until we receive a mouseup.

Bug: 1159951
Change-Id: I04f32745558885c6200f779570b6975ca1f4663a